nfc: move `MetadataBuilder` to `RootSignatureMetadata` library
