+++ llvm/lib/Frontend/HLSL/CMakeLists.txt
@@ -3,6 +3,7 @@ add_llvm_component_library(LLVMFrontendHLSL
   HLSLResource.cpp
   HLSLRootSignature.cpp
   HLSLRootSignatureUtils.cpp
+  RootSignatureMetadata.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/Frontend
