+++ llvm/lib/Frontend/HLSL/HLSLRootSignatureUtils.cpp
@@ -13,8 +13,6 @@
 #include "llvm/Frontend/HLSL/HLSLRootSignatureUtils.h"
 #include "llvm/ADT/SmallString.h"
 #include "llvm/ADT/bit.h"
-#include "llvm/IR/IRBuilder.h"
-#include "llvm/IR/Metadata.h"
 #include "llvm/Support/ScopedPrinter.h"
 
 namespace llvm {
@@ -210,8 +208,7 @@ raw_ostream &operator<<(raw_ostream &OS, const StaticSampler &Sampler) {
 namespace {
 
 // We use the OverloadVisit with std::visit to ensure the compiler catches if a
-// new RootElement variant type is added but it's operator<< or metadata
-// generation isn't handled.
+// new RootElement variant type is added but it's operator<< isn't handled.
 template <class... Ts> struct OverloadedVisit : Ts... {
   using Ts::operator()...;
 };
@@ -242,151 +239,6 @@ void dumpRootElements(raw_ostream &OS, ArrayRef<RootElement> Elements) {
     First = false;
   }
   OS << "}";
-}
-
-MDNode *MetadataBuilder::BuildRootSignature() {
-  const auto Visitor = OverloadedVisit{
-      [this](const dxbc::RootFlags &Flags) -> MDNode * {
-        return BuildRootFlags(Flags);
-      },
-      [this](const RootConstants &Constants) -> MDNode * {
-        return BuildRootConstants(Constants);
-      },
-      [this](const RootDescriptor &Descriptor) -> MDNode * {
-        return BuildRootDescriptor(Descriptor);
-      },
-      [this](const DescriptorTableClause &Clause) -> MDNode * {
-        return BuildDescriptorTableClause(Clause);
-      },
-      [this](const DescriptorTable &Table) -> MDNode * {
-        return BuildDescriptorTable(Table);
-      },
-      [this](const StaticSampler &Sampler) -> MDNode * {
-        return BuildStaticSampler(Sampler);
-      },
-  };
-
-  for (const RootElement &Element : Elements) {
-    MDNode *ElementMD = std::visit(Visitor, Element);
-    assert(ElementMD != nullptr &&
-           "Root Element must be initialized and validated");
-    GeneratedMetadata.push_back(ElementMD);
-  }
-
-  return MDNode::get(Ctx, GeneratedMetadata);
-}
-
-MDNode *MetadataBuilder::BuildRootFlags(const dxbc::RootFlags &Flags) {
-  IRBuilder<> Builder(Ctx);
-  Metadata *Operands[] = {
-      MDString::get(Ctx, "RootFlags"),
-      ConstantAsMetadata::get(Builder.getInt32(llvm::to_underlying(Flags))),
-  };
-  return MDNode::get(Ctx, Operands);
-}
-
-MDNode *MetadataBuilder::BuildRootConstants(const RootConstants &Constants) {
-  IRBuilder<> Builder(Ctx);
-  Metadata *Operands[] = {
-      MDString::get(Ctx, "RootConstants"),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Constants.Visibility))),
-      ConstantAsMetadata::get(Builder.getInt32(Constants.Reg.Number)),
-      ConstantAsMetadata::get(Builder.getInt32(Constants.Space)),
-      ConstantAsMetadata::get(Builder.getInt32(Constants.Num32BitConstants)),
-  };
-  return MDNode::get(Ctx, Operands);
-}
-
-MDNode *MetadataBuilder::BuildRootDescriptor(const RootDescriptor &Descriptor) {
-  IRBuilder<> Builder(Ctx);
-  std::optional<StringRef> TypeName =
-      getEnumName(dxil::ResourceClass(llvm::to_underlying(Descriptor.Type)),
-                  ArrayRef(ResourceClassNames));
-  assert(TypeName && "Provided an invalid Resource Class");
-  llvm::SmallString<7> Name({"Root", *TypeName});
-  Metadata *Operands[] = {
-      MDString::get(Ctx, Name),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Descriptor.Visibility))),
-      ConstantAsMetadata::get(Builder.getInt32(Descriptor.Reg.Number)),
-      ConstantAsMetadata::get(Builder.getInt32(Descriptor.Space)),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Descriptor.Flags))),
-  };
-  return MDNode::get(Ctx, Operands);
-}
-
-MDNode *MetadataBuilder::BuildDescriptorTable(const DescriptorTable &Table) {
-  IRBuilder<> Builder(Ctx);
-  SmallVector<Metadata *> TableOperands;
-  // Set the mandatory arguments
-  TableOperands.push_back(MDString::get(Ctx, "DescriptorTable"));
-  TableOperands.push_back(ConstantAsMetadata::get(
-      Builder.getInt32(llvm::to_underlying(Table.Visibility))));
-
-  // Remaining operands are references to the table's clauses. The in-memory
-  // representation of the Root Elements created from parsing will ensure that
-  // the previous N elements are the clauses for this table.
-  assert(Table.NumClauses <= GeneratedMetadata.size() &&
-         "Table expected all owned clauses to be generated already");
-  // So, add a refence to each clause to our operands
-  TableOperands.append(GeneratedMetadata.end() - Table.NumClauses,
-                       GeneratedMetadata.end());
-  // Then, remove those clauses from the general list of Root Elements
-  GeneratedMetadata.pop_back_n(Table.NumClauses);
-
-  return MDNode::get(Ctx, TableOperands);
-}
-
-MDNode *MetadataBuilder::BuildDescriptorTableClause(
-    const DescriptorTableClause &Clause) {
-  IRBuilder<> Builder(Ctx);
-  std::optional<StringRef> Name =
-      getEnumName(dxil::ResourceClass(llvm::to_underlying(Clause.Type)),
-                  ArrayRef(ResourceClassNames));
-  assert(Name && "Provided an invalid Resource Class");
-  Metadata *Operands[] = {
-      MDString::get(Ctx, *Name),
-      ConstantAsMetadata::get(Builder.getInt32(Clause.NumDescriptors)),
-      ConstantAsMetadata::get(Builder.getInt32(Clause.Reg.Number)),
-      ConstantAsMetadata::get(Builder.getInt32(Clause.Space)),
-      ConstantAsMetadata::get(Builder.getInt32(Clause.Offset)),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Clause.Flags))),
-  };
-  return MDNode::get(Ctx, Operands);
-}
-
-MDNode *MetadataBuilder::BuildStaticSampler(const StaticSampler &Sampler) {
-  IRBuilder<> Builder(Ctx);
-  Metadata *Operands[] = {
-      MDString::get(Ctx, "StaticSampler"),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Sampler.Filter))),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Sampler.AddressU))),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Sampler.AddressV))),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Sampler.AddressW))),
-      ConstantAsMetadata::get(llvm::ConstantFP::get(llvm::Type::getFloatTy(Ctx),
-                                                    Sampler.MipLODBias)),
-      ConstantAsMetadata::get(Builder.getInt32(Sampler.MaxAnisotropy)),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Sampler.CompFunc))),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Sampler.BorderColor))),
-      ConstantAsMetadata::get(
-          llvm::ConstantFP::get(llvm::Type::getFloatTy(Ctx), Sampler.MinLOD)),
-      ConstantAsMetadata::get(
-          llvm::ConstantFP::get(llvm::Type::getFloatTy(Ctx), Sampler.MaxLOD)),
-      ConstantAsMetadata::get(Builder.getInt32(Sampler.Reg.Number)),
-      ConstantAsMetadata::get(Builder.getInt32(Sampler.Space)),
-      ConstantAsMetadata::get(
-          Builder.getInt32(llvm::to_underlying(Sampler.Visibility))),
-  };
-  return MDNode::get(Ctx, Operands);
 }
 
 } // namespace rootsig
